requirements.dev.txt
flake8>=3.9.2,<3.10
django-extensions>=4.0
black==25.11.0
ipython==8.18.1
bpython==0.26
Werkzeug==3.1.5